+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 ## 2.4.0
 
 - Updated and improved styles for BU Landing Pages and color palettes.
+- Adds PHPCS configuration file.
+- Updates and prepares SQL queries per PHPCS.
 
 ## 2.3.15
 
@@ -62,7 +64,6 @@
 ## 2.3.7
 
 - Removes duplicate call to burf-base, which is a dependency of burf-theme.
-- Add responsive_html_class() for use on the `<html>` tag for class name output/filtering.
 
 ## 2.3.61